♻️ Refactor Artifact utils structure for better separation of concerns
- Move constants.ts from utils/ to Artifact/ root (sibling relationship)
- Split utils.ts into separate files: constants.ts and generateHeadingId.ts
- Update all import paths to reflect new file structure
- Maintain formatArtifactToMarkdown.ts in utils/ directory
- Add index.ts for utils exports

This creates clearer separation between constants and utility functions,
improving code organization and maintainability.
