+++ frontend/apps/app/components/SessionDetailPage/components/Output/components/Artifact/constants.ts
@@ -1,0 +1,5 @@
+export const EXECUTION_SECTION_TITLE = 'Execution History'
+export const SUCCESS_ICON = '✅'
+export const FAILURE_ICON = '❌'
+export const SUCCESS_STATUS = 'Success'
+export const FAILURE_STATUS = 'Failed'
