+++ frontend/apps/app/components/SessionDetailPage/components/Output/components/Artifact/utils/generateHeadingId.ts
@@ -1,0 +1,7 @@
+export const generateHeadingId = (text: string): string => {
+  return text
+    .toLowerCase()
+    .trim()
+    .replace(/[^\p{L}\p{N}_-]+/gu, '-')
+    .replace(/^-+|-+$/g, '')
+}
